Refactoring (should be easier to understand now) added some comments
